refactor: isolate validation functions in a separate file
